Move scan_serial_impl.h head file to 'support' folder.
