+++ test/parallel_api/numeric/numeric.ops/scan.pass.cpp
@@ -19,7 +19,7 @@
 #include _PSTL_TEST_HEADER(numeric)
 
 #include "support/utils.h"
-#include "scan_serial_impl.h"
+#include "support/scan_serial_impl.h"
 
 #if  !defined(_PSTL_TEST_INCLUSIVE_SCAN) && !defined(_PSTL_TEST_EXCLUSIVE_SCAN)
 #define _PSTL_TEST_INCLUSIVE_SCAN
